+++ src/cldfviz/__init__.py
@@ -1,4 +1,4 @@
 import pathlib
 
-__version__ = '1.2.1.dev0'
+__version__ = '1.3.0'
 PKG_DIR = pathlib.Path(__file__).parent