+++ CHANGELOG.md
@@ -1,7 +1,7 @@
 # Changes
 
 
-## [Unreleased]
+## [v1.3.0] - 2024-09-25
 
 - Use markdown support from `clldutils.source` if available.
 - Allow using code IDs in colormaps for `cldfviz.map`.